+++ test/form/samples/import-external-namespace-and-default/_expected/amd.js
@@ -1,6 +1,6 @@
 define(['foo'], function (foo) { 'use strict';
 
-	var foo__default = foo['default'];
+	var foo__default = 'default' in foo ? foo['default'] : foo;
 
 	console.log( foo.bar );