+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.59.0
 *unreleased*
 * Add `esm` format alias ([#2102](https://github.com/rollup/rollup/pull/2102))
+* Improve handling of external modules imported as both default and named imports ([#2136](https://github.com/rollup/rollup/pull/2136))
 * Fix wrong file name in error message ([#2137](https://github.com/rollup/rollup/pull/2137))
 
 ## 0.58.1